refactor test cases for BitVector
